+++ src/lib/adapters/database/redis/index.ts
@@ -2,7 +2,7 @@ import { DatabaseAdapter } from "@/lib/core/interfaces";
 import { RedisSchema } from "@/lib/adapters/definitions";
 import { dump } from "./dump";
 import { restore, prepareRestore } from "./restore";
-import { test, getDatabases } from "./connection";
+import { test, getDatabases, getDatabasesWithStats } from "./connection";
 import { analyzeDump } from "./analyze";
 import { getTables, getTableData } from "./browser";
 
@@ -17,6 +17,7 @@ export const RedisAdapter: DatabaseAdapter = {
     prepareRestore,
     test,
     getDatabases,
+    getDatabasesWithStats,
     analyzeDump,
     getTables,
     getTableData,